refact: splitting main file in multiple files under the same App namespace.
